+++ src/beignet/nn/__init__.py
@@ -1,26 +1,20 @@
 from ._adaptive_layer_norm import AdaptiveLayerNorm
-from ._alphafold3 import AlphaFold3
-from ._alphafold3_confidence import AlphaFold3Confidence
-from ._alphafold3_diffusion import AlphaFold3Diffusion
-from ._alphafold3_distogram import AlphaFold3Distogram
-from ._alphafold3_msa import AlphaFold3MSA
-from ._alphafold3_template_embedder import AlphaFold3TemplateEmbedder
-from ._atom_attention_decoder import AtomAttentionDecoder
+from ._alphafold3 import AlphaFold3, _Confidence, _Distogram, _InputFeatureEmbedder
 from ._atom_attention_encoder import AtomAttentionEncoder
 from ._atom_transformer import AtomTransformer
 from ._attention_pair_bias import AttentionPairBias
-from ._centre_random_augmentation import CentreRandomAugmentation
-from ._conditioned_transition_block import ConditionedTransitionBlock
-from ._diffusion_conditioning import DiffusionConditioning
-from ._diffusion_transformer import DiffusionTransformer
-from ._fourier_embedding import FourierEmbedding
-from ._input_feature_embedder import InputFeatureEmbedder
-from ._msa_pair_weighted_averaging import MSAPairWeightedAveraging
-from ._outer_product_mean import OuterProductMean
-from ._pairformer import Pairformer, PairformerBlock, SingleRowAttention
-from ._pairformer_stack import PairformerStack, PairformerStackBlock
+from ._diffusion_transformer import DiffusionTransformer, _ConditionedTransitionBlock
+from ._msa import MSA, _MSAPairWeightedAveraging, _OuterProductMean
+from ._pairformer_stack import PairformerStack, _PairformerStackBlock
 from ._relative_position_encoding import RelativePositionEncoding
-from ._sample_diffusion import SampleDiffusion
+from ._sample_diffusion import (
+    SampleDiffusion,
+    _AtomAttentionDecoder,
+    _CentreRandomAugmentation,
+    _Diffusion,
+    _FourierEmbedding,
+)
+from ._template_embedder import TemplateEmbedder
 from ._transition import Transition
 from ._triangle_attention_ending_node import TriangleAttentionEndingNode
 from ._triangle_attention_starting_node import TriangleAttentionStartingNode
@@ -31,28 +25,15 @@ __all__ = [
     "AdaptiveLayerNorm",
     "AlphaFold3",
     "AtomAttentionEncoder",
-    "AtomAttentionDecoder",
     "AtomTransformer",
     "AttentionPairBias",
-    "CentreRandomAugmentation",
-    "ConditionedTransitionBlock",
-    "AlphaFold3Confidence",
-    "DiffusionConditioning",
-    "AlphaFold3Diffusion",
     "DiffusionTransformer",
-    "AlphaFold3Distogram",
-    "FourierEmbedding",
-    "InputFeatureEmbedder",
-    "AlphaFold3MSA",
-    "MSAPairWeightedAveraging",
-    "OuterProductMean",
-    "Pairformer",
-    "PairformerBlock",
+    "MSA",
     "PairformerStack",
-    "PairformerStackBlock",
+    "_PairformerStackBlock",
     "RelativePositionEncoding",
     "SampleDiffusion",
-    "AlphaFold3TemplateEmbedder",
+    "TemplateEmbedder",
     "Transition",
     "TriangleAttentionEndingNode",
     "TriangleAttentionStartingNode",
